Create 10-Matching_socks.py
There is a large pile of socks that must be paired by color. Given an array of integers representing the color of each sock, determine how many pairs of socks with matching colors there are.

Example

ar = [1,2,1,2,1,3,2]

There is one pair of color 1 and one of color 2. There are three odd socks left, one of each color. The number of

pairs is 2. Complete the sockMerchant function and return

the number of pairs

Sample Input:

[10, 20, 20, 10, 10, 30, 50, 10, 20] Sample Output:

3

Explanation: (20~20) (10~10) (10~10) 30 50 20