Create List_Comprehesion.py
Let's learn about list comprehensions! You are given three integers  x, y and z representing the dimensions of a cuboid along with an integer n. Print a list of all possible coordinates given by (i, j, k) on a 3D grid where the sum of  is not equal to n.